+++ argilla-server/CHANGELOG.md
@@ -22,6 +22,7 @@ These are the section headers that we use:
 - Added tests
   - Response: update duration
   - Files: get, put, list, delete
+  - Models: get, post, put, delete
 
 ### "Changed"
 - Updated elasticsearch to 8.15.0